% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.express.dgC.R
\name{average.express.dgC}
\alias{average.express.dgC}
\title{Title This funciton calculate the average expression matrix from the sparse matrix.}
\usage{
average.express.dgC(mat, meta.data, ident.use, verbose = TRUE)
}
\arguments{
\item{verbose}{}

\item{object}{a Seurat object}

\item{ident}{the column that will be used as identities from the meta data}

\item{assay}{the assay that will be used}

\item{slot}{the slot that will be used}
}
\value{
An average expression matrix
}
\description{
Title This funciton calculate the average expression matrix from the sparse matrix.
}
\examples{
\dontrun{average.express(x = dgCMatrix,meta.data = meta.data,ident.use = "avtive.ident", verbose = TRUE)}
}
