% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.express.R
\name{average.express}
\alias{average.express}
\title{Title This funciton calculate the average expression matrix from the sparse matrix.}
\usage{
average.express(object, assay = "RNA", slot = "data", verbose = TRUE)
}
\arguments{
\item{object}{a Seurat object}

\item{assay}{the assay that will be used}

\item{slot}{the slot that will be used}

\item{verbose}{}

\item{ident}{the column that will be used as identities from the meta data}
}
\value{
An average expression matrix
}
\description{
Title This funciton calculate the average expression matrix from the sparse matrix.
}
\examples{
\dontrun{average.express(x = pbmc_small,assay = "RNA",genes = genes)}
}
