% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_pct.pos.big.R
\name{cal_pctpos.big}
\alias{cal_pctpos.big}
\title{cal_pctpos.big}
\usage{
cal_pctpos.big(x, ident.use, mat, thresh.min = 0)
}
\arguments{
\item{x}{the seurat object}

\item{ident.use}{the ident to use}

\item{mat}{marker gene table of seurat}

\item{thresh.min}{}
}
\value{
a modified marker gene matrix with a new colum named "pct.pos"
}
\description{
cal_pctpos.big
}
\examples{
\dontrun{cal_pctpos.big(x = x,ident.use = ident.use,mat = mat,thresh.min = 0)}
}
