% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exprs_markers.R
\docType{data}
\name{data_exprs_markers}
\alias{data_exprs_markers}
\title{data_exprs_markers example data}
\format{
A matrix with length(marker_genes) observations for each cluster.
}
\usage{
data_exprs_markers
}
\description{
This is an example data set calculated by starTracer::searchMarker().
This is a matrix of a Pseudobulkdata with all the marker genes found by seachMarker()
}
\keyword{datasets}
