% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_genes_markers.R
\docType{data}
\name{data_genes_markers}
\alias{data_genes_markers}
\title{data_genes_markers example data}
\format{
A vetor.
}
\usage{
data_genes_markers
}
\description{
This is a vector containing genes calculated by searchMarkers()
}
\keyword{datasets}
