% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMarker.R
\name{filterMarker}
\alias{filterMarker}
\title{filterMarker}
\usage{
filterMarker(x, ident.use, mat, num = 4)
}
\arguments{
\item{x}{Seurat Object}

\item{ident.use}{ident to use during calculation, should be the same as the one user used during FindAllMarkers()}

\item{mat}{the out out data.frame of FindAllMarkers()}

\item{num}{the number of marker genes the user prefers}
}
\value{
a modified data.frame of the FindAllMarkers function
}
\description{
filterMarker
}
\examples{
\dontrun{filterMarker(x = pbmc_small, ident.use = "RNA_snn_res.1", mat = diff.wilcox)}
}
