% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtGene_negative.R
\name{filtGene.Neg}
\alias{filtGene.Neg}
\title{filtGene_nagative}
\usage{
filtGene.Neg(expr.norm.neg, thresh.2.neg, clusters)
}
\arguments{
\item{expr.norm.neg}{the max.nomalized expression matrix for negative marker}

\item{thresh.2.neg}{this is the threshold range from 0 to 1. thresh.2.neg will be automatically mapped to the different quantiles of the range of the mean expression level in each different cluster}

\item{clusters}{a vector of all clusters in the single cell data}
}
\value{
a matrix fater filtering
}
\description{
this function is used to filter genes in each cluster acoording to the given thresh.2.neg when calculating negative markers
}
\examples{
\dontrun{filtGene(expr.norm.neg = x,thresh.2.neg = 0.5)}
}
