% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtGene.R
\name{filtGene}
\alias{filtGene}
\title{filtGene}
\usage{
filtGene(expr.norm, thresh.2, clusters)
}
\arguments{
\item{expr.norm}{the max.nomalized expression matrix}

\item{thresh.2}{this is the threshold range from 0 to 1. Thresh.2 will be automatically mapped to the different quantiles of the range of the mean expression level in each different cluster}

\item{clusters}{a vector of all clusters in the single cell data}
}
\value{
a matrix fater filtering
}
\description{
this function is used to filter genes in each cluster acoording to the given thresh.2
}
\examples{
\dontrun{filtGene(expr.norm = x,thresh.2 = 0.5)}
}
