% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatPlot.R
\name{HeatPlot}
\alias{HeatPlot}
\title{HeatPlot}
\usage{
HeatPlot(x, assay, genes)
}
\arguments{
\item{x}{a seurat object}

\item{assay}{the assay in the Seurat Object to use}

\item{genes}{the genes for the Seurat Object to use}
}
\value{
a plot of heatmap
}
\description{
HeatPlot
}
\examples{
\dontrun{HeatPlot(x = pbmc_small,assay = "RNA",genes = genes)}
}
