% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatPlotDgC.R
\name{HeatPlotDgC}
\alias{HeatPlotDgC}
\title{HeatPlotDgC}
\usage{
HeatPlotDgC(x, genes)
}
\arguments{
\item{x}{the average expression matrix}

\item{genes}{genes we will use}
}
\value{
a heatmap
}
\description{
HeatPlotDgC
}
\examples{
\dontrun{HeatPlotDgC(x = pbmc_small,genes = genes)}
}
