% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.pbmc_small.R
\docType{data}
\name{mat.pbmc_small}
\alias{mat.pbmc_small}
\title{mat.pbmc_small example data}
\format{
A matrix with 3 clusters named as "0", "1", "2":
}
\usage{
mat.pbmc_small
}
\description{
This is an example data set from pbmc, an average expression matrix
}
\keyword{datasets}
