% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PraraFrame.R
\name{ParaFrame}
\alias{ParaFrame}
\title{ParaFrame}
\usage{
ParaFrame(expr.norm, n.cluster, thresh.1, clusters)
}
\arguments{
\item{expr.norm}{max normalized function calculated before, but with a extra line of max.X}

\item{n.cluster}{cluster number}

\item{thresh.1}{threshold to tell high expression data from the low expression level}

\item{clusters}{clusters from the single cell experiment}
}
\value{
a dataframe of all the parameters
}
\description{
ParaFrame
}
\examples{
\dontrun{ParaFrame(expr.norm = expr.norm,n.cluster = n.cluster,thresh.1 = thresh.1)}
}
