% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmc_small.R
\docType{data}
\name{pbmc_small}
\alias{pbmc_small}
\title{pbmc_small example data}
\format{
An object of class \code{Seurat} with 230 rows and 80 columns.
}
\usage{
pbmc_small
}
\description{
This is an example data set from pbmc
}
\keyword{datasets}
