% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_kidney.R
\docType{data}
\name{sc_kidney}
\alias{sc_kidney}
\title{sc_kidney example data}
\format{
An object of class \code{Seurat} with 18323 rows and 16119 columns.
}
\usage{
sc_kidney
}
\description{
This is an example data set
}
\keyword{datasets}
