% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMarker.R
\name{searchMarker}
\alias{searchMarker}
\title{searchMarker}
\usage{
searchMarker(
  x,
  thresh.1 = 0.5,
  thresh.2 = NULL,
  method = "del_MI",
  num = 2,
  gene.use = NULL,
  meta.data = NULL,
  ident.use = NULL
)
}
\arguments{
\item{x}{a Seurat Object, an average expression matrix or a sparse matrix.}

\item{thresh.1}{the threshold for dividing clusters max-normalized gene expression to high or low, set to 0.5 by default}

\item{thresh.2}{the specificity level or the lowest expression level of the genes user prefers}

\item{method}{the arranging method of ordering the genes. Set default and strongly suggestted as "del_MI"}

\item{num}{the topN gene user prefers}

\item{gene.use}{the genes we use, set NULL to use all the genes or set as "HVG" to only use HVG in Seurat Object}

\item{meta.data}{when importing a dgCMatrix and a meta.data, users should pass the meta.data to this param}

\item{ident.use}{when importing a dgCMatrix and a meta.data, users should specify the ident that should be used in the meta.data}
}
\value{
a list containing a pseudobulk expression matrix of the selected genes; a gene list; a parameter frame and a heatmap
}
\description{
this is the main function of searching marker genes
}
\examples{
 \dontrun{calMarker(expr = expr,thresh.1 = 0.5, thresh.2 = 0.4, num = 2, method = "del_MI")}

}
