% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StVlnPlot.R
\name{StVlnPlot}
\alias{StVlnPlot}
\title{StVlnPlot}
\usage{
StVlnPlot(x, assay = "RNA", genes, ident = "seurat_clusters", colors = NULL)
}
\arguments{
\item{x}{a seurat object}

\item{assay}{the assay to use, set as "RNA" by default}

\item{genes}{genes to draw plot, we recommand to use the output gene list of starTracer::searchMarker()}

\item{ident}{the ident to use to draw a plot, the names must be a column from the seuratobject@metadata}

\item{colors}{colors to use}
}
\value{
a stackedVionlnPlot
}
\description{
StVlnPlot
}
\examples{
\dontrun{
StVlnPlot(x = pbmc_small,
assay = "RNA",
genes = genes,
ident = "RNA_snn_res.1",
colors = NULL)
}
}
