% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatPlot.R
\name{HeatPlot}
\alias{HeatPlot}
\title{HeatPlot}
\usage{
HeatPlot(
  x,
  genes,
  genes.top1,
  scale.method = "scale",
  lim.scale = 2,
  border_color = "black",
  clusters = clusters,
  colors = viridis(10)
)
}
\arguments{
\item{x}{a seurat object}

\item{genes}{the genes for the Seurat Object to use}

\item{scale.method}{the method use to scale the original matrix}

\item{lim.scale}{when using "scale" as the method of scale.method, scaled value will be limited from -lim.scale to lim.scale}

\item{border_color}{the color of border in the heatmap}

\item{clusters}{the clusters of import data}

\item{colors}{the colorkey of the heatmap}
}
\value{
a plot of heatmap
}
\description{
HeatPlot
}
\examples{
\dontrun{HeatPlot(x = pbmc_small,assay = "RNA",genes = genes)}
}
