% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMarker.R
\name{searchMarker}
\alias{searchMarker}
\title{searchMarker}
\usage{
searchMarker(
  x,
  thresh.1 = 0.5,
  thresh.2 = NULL,
  thresh.2.neg = NULL,
  method = "pos",
  num = 2,
  gene.use = NULL,
  meta.data = NULL,
  ident.use = NULL,
  scale.method = "scale",
  lim.scale = 2,
  border_color = "black",
  colors = viridis(10)
)
}
\arguments{
\item{x}{a Seurat Object, an average expression matrix or a sparse matrix.}

\item{thresh.1}{the threshold for dividing clusters max-normalized gene expression to high or low, set to 0.5 by default}

\item{thresh.2}{the the lowest expression level of the genes user prefers}

\item{thresh.2.neg}{the the highest expression level of the genes user prefers when calculating negative markers}

\item{method}{the arranging method of ordering the genes. Set default and strongly suggestted as "pos", starTracer will only find positivie marker genes, setting as "all" to out put both up-regulated and down-regulated markers}

\item{num}{the topN gene user prefers}

\item{gene.use}{the genes we use, set NULL to use all the genes or set as "HVG" to only use HVG in Seurat Object}

\item{meta.data}{when importing a dgCMatrix and a meta.data, users should pass the meta.data to this param}

\item{ident.use}{when importing a dgCMatrix and a meta.data, users should specify the ident that should be used in the meta.data}

\item{scale.method}{the scaling method that will be used when outputting the heatmap}

\item{lim.scale}{when using "scale" as the method of scale.method, scaled value will be limited from -lim.scale to lim.scale}

\item{border_color}{the color of border in the heatmap}

\item{colors}{the colorkey of the heatmap}
}
\value{
a list containing a pseudobulk expression matrix of the selected genes; a gene list; a parameter frame and a heatmap
}
\description{
this is the main function of searching marker genes
}
\examples{
 \dontrun{calMarker(expr = expr,thresh.1 = 0.5, thresh.2 = 0.4, num = 2, method = "pos")}

}
