% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMarkers.R
\name{searchMarker}
\alias{searchMarker}
\title{searchMarker}
\usage{
searchMarker(x, thresh.1 = 0.3, thresh.2 = 0.1, method = "del_MI", num = 2)
}
\arguments{
\item{x}{the psudobulk RNA expression matrix of a Seurat Object}

\item{thresh.1}{the threshold for dividing clusters max-normalized gene expression to high or low, set to 0.5 by default}

\item{thresh.2}{the specificity level or the lowest expression level of the genes user prefers}

\item{method}{the arranging method of ordering the genes. Set default and strongly suggestted as "del_MI"}

\item{num}{the topN gene user prefers}
}
\value{
a list containing a pseudobulk expression matrix of the selected genes; a gene list; a parameter frame.
}
\description{
this is the main function of searching marker genes
}
\examples{
 \dontrun{calMarker(expr = expr,thresh.1 = 0.5, thresh.2 = 0.4, num = 2, method = "del_MI")}

}
